requirements.txt
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
borb==2.1.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
cssselect2==0.7.0
deprecation==2.1.0
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-querycount==0.8.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
fonttools==4.38.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==4.12.0
lxml==4.9.1
Markdown==3.4.1
numpy==1.24.2
openpyxl==3.1.1
oscrypto==1.3.0
packaging==21.3
pandas==1.5.3
pdf2image==1.16.0
pikepdf==6.2.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
PyJWT==2.5.0
PyMuPDF==1.20.2
pyparsing==3.0.9
pypdf==3.5.0
python-barcode==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.12
webencodings==0.5.1
XlsxWriter==3.0.3
zipp==3.8.1
